requirements.txt
telethon==1.28.5
openai==0.27.8
python-dotenv==1.0.0